Feature/chapter 11 supabaseの反映

+++ src/app/admin/posts/[id]/page.tsx
@@ -3,27 +3,31 @@ import React, { useEffect, useState } from "react";
 import { useParams, useRouter } from "next/navigation";
 import { Category } from "@/app/_types/Category";
 import PostForm from "../_components/PostForm";
+import { useSupabaseSession } from "@/app/hooks/useSupabaseSession";
+import { log } from "console";
 
 const PostPage = () => {
   const [title, setTitle] = useState("");
   const [content, setContent] = useState("");
-  const [thumbnailUrl, setThumbnailUrl] = useState("");
+  const [thumbnailImageKey, setThumbnailImageKey] = useState("");
   const [categories, setCategories] = useState<Category[]>([]);
   const { id } = useParams();
   const router = useRouter();
   const [selectedCategories, setSelectedCategories] = useState<Category[]>([]);
 
+  const { token } = useSupabaseSession(); //token追加
+
   const handlePostSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-
-    //記事更新
+    if (!token) return;
+    //記事更新　画像をアップロードしてthumbnailImageKeyを保存する処理
     await fetch(`/api/admin/posts/${id}`, {
       method: "PUT",
-      headers: { "Content-type": "application/json" },
+      headers: { "Content-type": "application/json", Authorization: token },
       body: JSON.stringify({
         title,
         content,
-        thumbnailUrl,
+        thumbnailImageKey,
         categories: selectedCategories,
       }),
     });
@@ -34,32 +38,37 @@ const PostPage = () => {
   //記事削除
   const handleDeletePost = async () => {
     if (!confirm("記事を削除しますか？")) return;
+    if (!token) return;
     await fetch(`/api/admin/posts/${id}`, {
       method: "DELETE",
+      headers: { "Content-Type": "application/json", Authorization: token },
     });
     alert("記事を削除しました。");
     router.push("/admin/posts");
   };
 
-  //記事を取得
+  //記事更新：画像をアップロード及びthumbnailImageKeyを保存した後の記事取得
   useEffect(() => {
+    if (!token) return;
     const fetcher = async () => {
-      const res = await fetch(`/api/admin/posts/${id}`);
+      const res = await fetch(`/api/admin/posts/${id}`, {
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
+      });
       // const data = await res.json();
       // console.log(data);
 
       const { post } = await res.json();
       setTitle(post.title);
       setContent(post.content);
-      setThumbnailUrl(post.thumbnailUrl);
+      setThumbnailImageKey(post.thumbnailImageKey);
 
       // カテゴリーの状態を更新↓
       const postCategories = post.postCategories;
       // 中間テーブルをはさんでいる。console.logで何がはいっているか確認し、設定
       // ・・post.postCategoriesの中のカテゴリーオブジェクトを取ってきている。
-
-      console.log(post);
-      console.log(postCategories);
       const categories: Category[] = postCategories.map(
         (pc: { category: Category }) => pc.category
       );
@@ -69,23 +78,24 @@ const PostPage = () => {
       // ));
     };
     fetcher();
-  }, [id]);
+  }, [id, token]);
 
   //カテゴリーを取得
   //  useEffect(()=>{});
   useEffect(() => {
+    if (!token) return;
     const fetcher = async () => {
-      const res = await fetch("/api/admin/categories");
+      const res = await fetch("/api/admin/categories", {
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
+      });
       const { categories }: { categories: Category[] } = await res.json();
       setCategories(categories);
     };
     fetcher();
-  }, []);
-
-  // const handleSelectCategory = (category: Category) => {
-  //   setSelectedCategories([...selectedCategories, category]);
-  // };
-  // console.log(selectedCategories);
+  }, [token]);
 
   const handleSelectCategory = (category: Category) => {
     if (selectedCategories.map((c) => c.id).includes(category.id)) {
@@ -109,8 +119,8 @@ const PostPage = () => {
         setTitle={setTitle} // タイトルを更新する関数
         content={content} // 内容の状態
         setContent={setContent} // 内容を更新する関数
-        thumbnailUrl={thumbnailUrl} // サムネイルURLの状態
-        setThumbnailUrl={setThumbnailUrl} // サムネイルURLを更新する関数
+        thumbnailImageKey={thumbnailImageKey} // サムネイルURLの状態
+        setThumbnailImageKey={setThumbnailImageKey} // サムネイルURLを更新する関数
         categories={categories} // カテゴリーのリスト
         selectedCategories={selectedCategories} // 選択されたカテゴリーのリスト
         handleSelectCategory={handleSelectCategory} // カテゴリーを選択/解除する処理
